+++ CVarsManager.lua
@@ -278,7 +278,7 @@ local function SetCVarHook(name, value, c)
     local invalid = CVars.InvalidCVarsForOcclusionDetection()
     -- Update Hiding Nameplates only if something changed
     if (enabled and invalid) or (not enabled and not invalid) then
-      Addon.Transparency:UpdateSettings()
+      Addon.Transparency.UpdateSettings()
       --Addon:ForceUpdate()
     elseif invalid then
       Addon.Logging.Warning(L["Transparency for occluded units is being disabled as certain console variables (CVars) related to nameplate transparency are set in a way to prevent this feature from working."])