+++ Widgets/AurasWidget.lua
@@ -3309,10 +3309,10 @@ end
 local function UpdateWidgetTimeBarModeNoDuration(self, aura_frame, expiration, duration)
   if duration == 0 then
     aura_frame.Statusbar:SetValue(100)
-    Animation:StopFlash(aura_frame)
+    AnimationStopFlash(aura_frame)
   elseif expiration == 0 then
     aura_frame.Statusbar:SetValue(0)
-    Animation:StopFlash(aura_frame)
+    AnimationStopFlash(aura_frame)
   else
     local timeleft = expiration - GetTime()
     if timeleft > 60 then
@@ -3323,7 +3323,7 @@ local function UpdateWidgetTimeBarModeNoDuration(self, aura_frame, expiration, d
 
     local db = self.db_widget
     if db.FlashWhenExpiring and timeleft < db.FlashTime then
-      Animation:Flash(aura_frame)
+      AnimationFlash(aura_frame)
     end
 
     aura_frame.Statusbar:SetValue(timeleft * 100 / duration)